cloud_functions/common/requirements.txt
# =============================================================================
# DEPENDENCIAS COMUNES PARA CLOUD FUNCTIONS
# =============================================================================

# Dependencias del proyecto principal
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# Google Cloud dependencies
google-cloud-storage>=2.0.0
google-cloud-logging>=3.5.0

# Utilidades para manejo de datos
tenacity>=8.2.0 